+++ dart/constraint/JointCoulombFrictionConstraint.cpp
@@ -88,13 +88,6 @@ void JointCoulombFrictionConstraint::setConstraintForceMixing(double _cfm)
            << "It is set to 1e-9." << std::endl;
     mConstraintForceMixing = 1e-9;
   }
-  if (_cfm > 1.0)
-  {
-    dtwarn << "Constraint force mixing parameter[" << _cfm
-           << "] is greater than 1.0. "
-           << "It is set to 1.0." << std::endl;
-    mConstraintForceMixing = 1.0;
-  }
 
   mConstraintForceMixing = _cfm;
 }
@@ -118,7 +111,7 @@ void JointCoulombFrictionConstraint::update()
 
     if (mNegativeVel[i] != 0.0)
     {
-      double timeStep = mJoint->getSkeleton()->getTimeStep();
+      double timeStep = mJoint->getRawSkeleton()->getTimeStep();
       // TODO: There are multiple ways to get time step (or its inverse).
       //   - ContactConstraint get it from the constructor parameter
       //   - Skeleton has it itself.
@@ -183,7 +176,7 @@ void JointCoulombFrictionConstraint::applyUnitImpulse(std::size_t _index)
   assert(_index < mDim && "Invalid Index.");
 
   std::size_t localIndex = 0;
-  const dynamics::SkeletonPtr& skeleton = mJoint->getSkeleton();
+  dynamics::Skeleton* const skeleton = mJoint->getRawSkeleton();
 
   std::size_t dof = mJoint->getNumDofs();
   for (std::size_t i = 0; i < dof; ++i)
@@ -219,7 +212,7 @@ void JointCoulombFrictionConstraint::getVelocityChange(
     if (mActive[i] == false)
       continue;
 
-    if (mJoint->getSkeleton()->isImpulseApplied())
+    if (mJoint->getRawSkeleton()->isImpulseApplied())
       _delVel[localIndex] = mJoint->getVelocityChange(i);
     else
       _delVel[localIndex] = 0.0;
@@ -241,13 +234,13 @@ void JointCoulombFrictionConstraint::getVelocityChange(
 //==============================================================================
 void JointCoulombFrictionConstraint::excite()
 {
-  mJoint->getSkeleton()->setImpulseApplied(true);
+  mJoint->getRawSkeleton()->setImpulseApplied(true);
 }
 
 //==============================================================================
 void JointCoulombFrictionConstraint::unexcite()
 {
-  mJoint->getSkeleton()->setImpulseApplied(false);
+  mJoint->getRawSkeleton()->setImpulseApplied(false);
 }
 
 //==============================================================================
@@ -272,7 +265,7 @@ void JointCoulombFrictionConstraint::applyImpulse(double* _lambda)
 //==============================================================================
 dynamics::SkeletonPtr JointCoulombFrictionConstraint::getRootSkeleton() const
 {
-  return mJoint->getSkeleton()->mUnionRootSkeleton.lock();
+  return mJoint->getRawSkeleton()->mUnionRootSkeleton.lock();
 }
 
 //==============================================================================