+++ dart/constraint/WeldJointConstraint.cpp
@@ -196,15 +196,15 @@ void WeldJointConstraint::applyUnitImpulse(std::size_t _index)
     assert(mBodyNode1->isReactive() || mBodyNode2->isReactive());
 
     // Weld joint between two bodies in one skeleton
-    if (mBodyNode1->getSkeleton() == mBodyNode2->getSkeleton())
+    if (mBodyNode1->getRawSkeleton() == mBodyNode2->getRawSkeleton())
     {
-      mBodyNode1->getSkeleton()->clearConstraintImpulses();
+      mBodyNode1->getRawSkeleton()->clearConstraintImpulses();
 
       if (mBodyNode1->isReactive())
       {
         if (mBodyNode2->isReactive())
         {
-          mBodyNode1->getSkeleton()->updateBiasImpulse(
+          mBodyNode1->getRawSkeleton()->updateBiasImpulse(
               mBodyNode1,
               mJacobian1.row(_index),
               mBodyNode2,
@@ -212,7 +212,7 @@ void WeldJointConstraint::applyUnitImpulse(std::size_t _index)
         }
         else
         {
-          mBodyNode1->getSkeleton()->updateBiasImpulse(
+          mBodyNode1->getRawSkeleton()->updateBiasImpulse(
               mBodyNode1, mJacobian1.row(_index));
         }
       }
@@ -220,7 +220,7 @@ void WeldJointConstraint::applyUnitImpulse(std::size_t _index)
       {
         if (mBodyNode2->isReactive())
         {
-          mBodyNode2->getSkeleton()->updateBiasImpulse(
+          mBodyNode2->getRawSkeleton()->updateBiasImpulse(
               mBodyNode2, -mJacobian2.row(_index));
         }
         else
@@ -229,25 +229,25 @@ void WeldJointConstraint::applyUnitImpulse(std::size_t _index)
         }
       }
 
-      mBodyNode1->getSkeleton()->updateVelocityChange();
+      mBodyNode1->getRawSkeleton()->updateVelocityChange();
     }
     // Weld joint between two bodies in different skeletons
     else
     {
       if (mBodyNode1->isReactive())
       {
-        mBodyNode1->getSkeleton()->clearConstraintImpulses();
-        mBodyNode1->getSkeleton()->updateBiasImpulse(
+        mBodyNode1->getRawSkeleton()->clearConstraintImpulses();
+        mBodyNode1->getRawSkeleton()->updateBiasImpulse(
             mBodyNode1, mJacobian1.row(_index));
-        mBodyNode1->getSkeleton()->updateVelocityChange();
+        mBodyNode1->getRawSkeleton()->updateVelocityChange();
       }
 
       if (mBodyNode2->isReactive())
       {
-        mBodyNode2->getSkeleton()->clearConstraintImpulses();
-        mBodyNode2->getSkeleton()->updateBiasImpulse(
+        mBodyNode2->getRawSkeleton()->clearConstraintImpulses();
+        mBodyNode2->getRawSkeleton()->updateBiasImpulse(
             mBodyNode2, -mJacobian2.row(_index));
-        mBodyNode2->getSkeleton()->updateVelocityChange();
+        mBodyNode2->getRawSkeleton()->updateVelocityChange();
       }
     }
   }
@@ -255,10 +255,10 @@ void WeldJointConstraint::applyUnitImpulse(std::size_t _index)
   {
     assert(mBodyNode1->isReactive());
 
-    mBodyNode1->getSkeleton()->clearConstraintImpulses();
-    mBodyNode1->getSkeleton()->updateBiasImpulse(
+    mBodyNode1->getRawSkeleton()->clearConstraintImpulses();
+    mBodyNode1->getRawSkeleton()->updateBiasImpulse(
         mBodyNode1, mJacobian1.row(_index));
-    mBodyNode1->getSkeleton()->updateVelocityChange();
+    mBodyNode1->getRawSkeleton()->updateVelocityChange();
   }
 
   mAppliedImpulseIndex = _index;
@@ -271,12 +271,12 @@ void WeldJointConstraint::getVelocityChange(double* _vel, bool _withCfm)
   assert(isActive());
 
   Eigen::Vector6d velChange = Eigen::Vector6d::Zero();
-  if (mBodyNode1->getSkeleton()->isImpulseApplied() && mBodyNode1->isReactive())
+  if (mBodyNode1->getRawSkeleton()->isImpulseApplied() && mBodyNode1->isReactive())
   {
     velChange += mBodyNode1->getBodyVelocityChange();
   }
 
-  if (mBodyNode2 && mBodyNode2->getSkeleton()->isImpulseApplied()
+  if (mBodyNode2 && mBodyNode2->getRawSkeleton()->isImpulseApplied()
       && mBodyNode2->isReactive())
   {
     velChange -= mJacobian2 * mBodyNode2->getBodyVelocityChange();
@@ -298,26 +298,26 @@ void WeldJointConstraint::getVelocityChange(double* _vel, bool _withCfm)
 void WeldJointConstraint::excite()
 {
   if (mBodyNode1->isReactive())
-    mBodyNode1->getSkeleton()->setImpulseApplied(true);
+    mBodyNode1->getRawSkeleton()->setImpulseApplied(true);
 
   if (mBodyNode2 == nullptr)
     return;
 
   if (mBodyNode2->isReactive())
-    mBodyNode2->getSkeleton()->setImpulseApplied(true);
+    mBodyNode2->getRawSkeleton()->setImpulseApplied(true);
 }
 
 //==============================================================================
 void WeldJointConstraint::unexcite()
 {
   if (mBodyNode1->isReactive())
-    mBodyNode1->getSkeleton()->setImpulseApplied(false);
+    mBodyNode1->getRawSkeleton()->setImpulseApplied(false);
 
   if (mBodyNode2 == nullptr)
     return;
 
   if (mBodyNode2->isReactive())
-    mBodyNode2->getSkeleton()->setImpulseApplied(false);
+    mBodyNode2->getRawSkeleton()->setImpulseApplied(false);
 }
 
 //==============================================================================
@@ -343,13 +343,13 @@ void WeldJointConstraint::applyImpulse(double* _lambda)
 dynamics::SkeletonPtr WeldJointConstraint::getRootSkeleton() const
 {
   if (mBodyNode1->isReactive())
-    return mBodyNode1->getSkeleton()->mUnionRootSkeleton.lock();
+    return mBodyNode1->getRawSkeleton()->mUnionRootSkeleton.lock();
 
   if (mBodyNode2)
   {
     if (mBodyNode2->isReactive())
     {
-      return mBodyNode2->getSkeleton()->mUnionRootSkeleton.lock();
+      return mBodyNode2->getRawSkeleton()->mUnionRootSkeleton.lock();
     }
     else
     {
@@ -373,7 +373,7 @@ void WeldJointConstraint::uniteSkeletons()
   if (!mBodyNode1->isReactive() || !mBodyNode2->isReactive())
     return;
 
-  if (mBodyNode1->getSkeleton() == mBodyNode2->getSkeleton())
+  if (mBodyNode1->getRawSkeleton() == mBodyNode2->getRawSkeleton())
     return;
 
   const dynamics::SkeletonPtr& unionId1