+++ dart/dynamics/Skeleton.hpp
@@ -1329,6 +1329,8 @@ public:
   ///
   std::weak_ptr<Skeleton> mUnionRootSkeleton;
 
+  Skeleton* mRawUnionRootSkeleton;
+
   ///
   std::size_t mUnionSize;