+++ CHANGELOG.md
@@ -25,6 +25,12 @@
   * Fixed compiler warnings from GCC 9.1: [#1366](https://github.com/dartsim/dart/pull/1366)
   * Replace M_PI with dart::math::constantsd::pi(): [#1367](https://github.com/dartsim/dart/pull/1367)
 
+### [DART 6.9.2 (2019-07-XX)](https://github.com/dartsim/dart/milestone/60?closed=1)
+
+* Dynamics
+
+  * Allowed constraint force mixing > 1: [#1371](https://github.com/dartsim/dart/pull/1371)
+
 ### [DART 6.9.1 (2019-06-06)](https://github.com/dartsim/dart/milestone/59?closed=1)
 
 * Collision